Delete, layers (WIP), set background image (WIP)
Elements can be deleted from the map and database with the Del key. The landmark layer has been integrated into the project, and buttons in the toolbar swap layers. The buttons do not set the other layers' objects to unselectable yet. The background can be set from the image bank, but number of tiles can't be chosen yet.